backend/pyproject.toml
[project]
name = "digital-co-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "markitdown[docx,pdf,pptx]>=0.1.2",
    "pymupdf4llm>=0.0.25",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.8.2",
    "pandas>=2.3.0",
    "pypdf2>=3.0.1",
    "flashrank>=0.2.10",
    "qdrant-client>=1.14.3",
    "sentence-transformers>=5.0.0",
    "pydantic-settings>=2.10.1",
    "einops>=0.8.1",
    "pdfplumber>=0.11.7",
    "tabulate>=0.9.0",
    "uvicorn>=0.35.0",
    "fastapi>=0.115.14",
    "python-multipart>=0.0.20",
    "black>=25.1.0",
    "isort>=6.0.1",
    "docling>=2.40.0",
    "boto3>=1.39.4",
    "tempmail-python>=2.3.3",
    "tempmail-lol>=3.0.0",
    "temp-mail>=0.2",
    "pymongo>=4.13.2",
    "git-filter-repo>=2.47.0",
    "newsapi-python>=0.2.7",
    "maturin>=1.9.1",
    "soundfile>=0.13.1",
    "openai>=1.93.0",
]
